requirements.txt
# Traffic Analyser dependencies
# Install with: pip install -r requirements.txt

opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0        # YOLOv8
mysql-connector-python>=8.0.0
